Create user microservice
