+++ backend/apps/userSvc/src/users/users.module.ts
@@ -1,0 +1,13 @@
+import { Logger, Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+
+import { UserModel } from './user.model';
+import { UsersController } from './users.controller';
+import { UsersService } from './users.service';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([UserModel])],
+  providers: [UsersService, Logger],
+  controllers: [UsersController],
+})
+export class UsersModule {}
