+++ backend/apps/userSvc/test/app.e2e-spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
-import { UserModule } from '../src/user.module';
+import { UserModule } from '../src/app.module';
 
 describe('UserController (e2e)', () => {
   let app: INestApplication;
